Organized routes into files, and fixed bug unsetting name taken when the nft is updated after the name.
